requirements.txt
# --- Backend Core ---
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
python-multipart

# --- Database (MySQL) ---
sqlalchemy>=2.0.0
mysql-connector-python>=8.3.0
pymysql

# --- AI & Vector DB ---
qdrant-client>=1.7.0
langchain>=0.1.0
langchain-community
langchain-core

# Các provider cụ thể cho LangChain (Quan trọng)
langchain-google-genai      # Dùng cho Chatbot chính (Gemini)
langchain-openai            # Dùng cho Data Generator (gpt-4o-mini)
langchain-huggingface       # Dùng cho Embedding (import_local.py)
langchain-qdrant            # Kết nối Langchain với Qdrant

# --- Xử lý Model & Embedding ---
transformers>=4.37.0
sentence-transformers
numpy

# --- Scraper Tools (Cào dữ liệu) ---
playwright                  # Dùng để cào Thư Viện Pháp Luật
beautifulsoup4              # Hỗ trợ xử lý HTML nếu cần

# --- Torch (Bản CPU cho nhẹ máy) ---
# Lưu ý: Nếu cài bị lỗi, hãy chạy lệnh cài riêng bên dưới
torch --index-url https://download.pytorch.org/whl/cpu